chore: Move jobs icon with number to jobs action icon
